Quality size sliders are implemented. Limits are calculated based on MB/Minute.

+++ NzbDrone.Core/Datastore/Migrations/Migration20110909.cs
@@ -17,8 +17,8 @@ namespace NzbDrone.Core.Datastore.Migrations
                                             {
                                                 new Column("QualityTypeId", DbType.Int32, ColumnProperty.PrimaryKey),
                                                 new Column("Name", DbType.String, ColumnProperty.NotNull),
-                                                new Column("MinSize", DbType.Int64, ColumnProperty.NotNull),
-                                                new Column("MaxSize", DbType.Int64, ColumnProperty.NotNull)
+                                                new Column("MinSize", DbType.Int32, ColumnProperty.NotNull),
+                                                new Column("MaxSize", DbType.Int32, ColumnProperty.NotNull)
                                             });
         }
